+++ Sources/NexusCoreUser/Network/CoreUserAPI.swift
@@ -55,7 +55,7 @@ final class CoreUserAPI: @unchecked Sendable {
             phone: JSONObject.string(data, keys: "phone"),
             emailBound: JSONObject.bool(data, key: "email_bound"),
             phoneBound: JSONObject.bool(data, key: "phone_bound"),
-            balance: JSONObject.int(data, key: "balance") ?? 0,
+            balance: JSONObject.double(data, key: "balance") ?? 0,
             userInfoSynced: true
         )
     }
